--- python files/requirments.txt
@@ -1,34 +1,0 @@
-PIL
-spacy
-google.generativeai
-opencv-python
-easyocr
-time
-torch
-torch.hub
-omegaconf
-pydub
-simpleaudio
-numpy
-os
-cvzone
-math
-requests
-time
-logging
-torchaudio
-transformers
-torch
-pyaudio
-face-recognition
-ultralytics
-multiprocessing
-requests
-haversine
-serial
-time
-string
-pynmea2
-firebase-admin
-datetime
-mpu6050
